// src/app/shared/model/srvy-task-obj.model.ts
export class SrvyTaskObj {
    SrvyTaskId: number;
    SrvyOrderId: number;
    SrvyTaskNo: string;
    MrSurveyTypeCode: string;
    MrSurveyTaskStatCode: string;
    MrSrvyObjTypeCode: string;
    RefNo: string;
    SurveyorId: number;
    RetrieveDt: Date;
    ResultDt: Date;
    ReviewByRefUserId: number;
    ReviewDt: Date;
    ReviewNotes: string;
    IsAddtReq: boolean;
    PrevSurveyTaskNo: string;
    Result: string;
    CustNo: string;
    MrCustModelCode: string;
    CustName: string;
    CustAddr: string;
    CustPhone: string;
    Zipcode: string;
    AssignDt: Date;
    SrvyFormSchmId: number;
    Notes: string;
    RefReasonId: number;
    PrevSurveyorId: number;
    MobileAssignmentId: number;
    Kelurahan: string;
    Kecamatan: string;
    RT: string;
    RW: string;
    City: string;
    Addr: string;
    RowVersion: string;


    constructor() {

    }
}